add fingerprint
